pyproject.toml
[project]
name = "toxic-classify"
version = "0.1.0"
description = "toxic-classify"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0",
    "transformers>=4.40",
    "torch>=2.0",
    "grpcio>=1.60",
    "protobuf>=4.25",
    "prometheus-client>=0.20",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.coverage.run]
source = ["src"]
omit = ["src/inference_pb2.py", "src/inference_pb2_grpc.py"]

[tool.isort]
profile = "black"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "grpcio-tools>=1.60",
    "black>=24.0",
    "isort>=5.13",
]
